backend/requirements.txt
alembic==1.7.3
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
zipp==3.6.0
